refector: useWindowSize 훅을 리팩토링
- window 객체를 사용하는 모든 컴포넌트를 SSR로 부르지 못하고 동적 로딩하는 것은 불편하다고 판단하여 window 객체의 유무로 분기하는 것으로 변경
- 사용하지 않는 innerHeight 제거
